move symbol table to data structures
